Week 9 - Day 2~3 Commit

+++ controller/CartController.js
@@ -1,15 +1,14 @@
+const ensureAuthorization = require('../auth'); // 인증모듈
 const jwt = require('jsonwebtoken');
 const conn = require('../mariadb');
 const { StatusCodes } = require('http-status-codes');
-const dotenv = require('dotenv');
-dotenv.config();
 
 // 장바구니 담기
 const addToCart = (req, res) => {
     const { book_id, quantity } = req.body;
 
     let authorization = ensureAuthorization(req, res);
-    
+
     if (authorization instanceof jwt.TokenExpiredError) {
         return res.status(StatusCodes.UNAUTHORIZED).json({
             "message": "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
@@ -45,12 +44,16 @@ const getCartItems = (req, res) => {
         });
     }
     else {
-
         let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price FROM cartItems 
         LEFT JOIN books 
         ON cartItems.book_id = books.id
-        WHERE user_id=? AND cartItems.id IN (?)`;
-        let values = [authorization.id, selected];
+        WHERE user_id=?`;
+        let values = [authorization.id];
+        if (selected) { // 주문서 작성 시 '선택한 장바구니 목록 조회'
+            sql += ` AND cartItems.id IN (?)`;
+            values.push(selected);
+        }
+
         conn.query(sql, values, function (err, results) {
             if (err) {
                 console.log(err);
@@ -63,38 +66,32 @@ const getCartItems = (req, res) => {
 };
 // 장바구니 도서 삭제
 const removeCartItem = (req, res) => {
-    // 좋아요 삭제
-    const cartItemsId = req.params.id;
-    let sql = "DELETE FROM cartItems WHERE id = ?";
-    conn.query(sql, cartItemsId, function (err, results) {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+    let authorization = ensureAuthorization(req, res);
 
-        res.status(StatusCodes.OK).json(results);
-
-    })
-};
-
-function ensureAuthorization(req, res) {
-    try {
-        let receivedJwt = req.headers["authorization"];
-        console.log("received jwt : ", receivedJwt);
-
-        let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-        console.log(decodedJwt);
-
-        return decodedJwt;
-
-    } catch (err) {
-        console.log(err.name);
-        console.log(err.message);
-
-
-        return err;
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        });
     }
-}
+    else {
+
+        const cartItemsId = req.params.id;
+        let sql = "DELETE FROM cartItems WHERE id = ?";
+        conn.query(sql, cartItemsId, function (err, results) {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+
+            res.status(StatusCodes.OK).json(results);
+
+        })
+    }
+};
 
 module.exports = {
     addToCart,
